Add files via upload
This Drowsiness Detection System uses OpenCV and Dlib to monitor a person’s eye state in real time. It detects facial landmarks and calculates the Eye Aspect Ratio (EAR) to determine if the eyes are closed. If EAR falls below a threshold , an alert message is displayed, and an alarm sound is played to wake the user. The system continuously processes frames from the webcam for real-time monitoring. It is useful for preventing accidents due to driver fatigue or prolonged screen exposure.